assistants/km-agno/pyproject.toml
[project]
name = "km-agno"
version = "0.1.0"
description = "Index flink-studies for semantic search with Agno and Chroma"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno[chromadb,openai]>=2.0.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
km-agno = "km_agno.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["km_agno"]

